+++ README.md
@@ -88,7 +88,7 @@
 * 단위 테스트하기 어려운 코드와 `단위 테스트 가능한 코드`를 `분리`하고, `단위 테스트 가능한 코드`에 대해 `단위 테스트를 구현`한다
 * JPA 구현체인 Spring Data Jpa로의 코드 마이그레이션을 수행한다
 * 현재 테스트의 보호를 받고 있는 상태로, 테스트의 보호 속에서 4-step 으로 리펙터링을 진행한다
-    1. `Application Layer`에 존재하는 비즈니스 로직(생성과 같은 단순한 기능)을 `Domain Layer`로 옮기기 (소극적 기능 리펙터링)(TDD)
+    1. `table.Application Layer`에 존재하는 비즈니스 로직(생성과 같은 단순한 기능)을 `Domain Layer`로 옮기기 (소극적 기능 리펙터링)(TDD)
     2. `Domain Layer` 와 `Infra Layer` 간의 의존 구조 개선하기
         * JPA 구현체인 Spring-Data-Jpa 도입
     3. `Infra Layer`의 보호 아래 `도메인 객체`를 세부적으로 분리 (적극적 기능 리펙터링)
@@ -106,7 +106,7 @@
 ## 요구사항
 
 * `메뉴` 정보가 변경 되더라도 `주문 항목`이 변경되지 않게 구현한다
-  * 2단계 - 의존성 리펙터링 리뷰 중 간접 참조하는 방식으로 개선 완료
+    * 2단계 - 의존성 리펙터링 리뷰 중 간접 참조하는 방식으로 개선 완료
 * 클래스 사이, 패키지 사이의 `의존 관계`는 `단방향`이 되도록 개선한다
 * DB 스키마 변경이 필요하면 Flyway 로 관리한다
 
@@ -168,7 +168,7 @@
     * 메뉴 그룹
     * 주문 테이블
 * 다른 서비스에 의존적인 도메인
-    * 주문  
+    * 주문
         * 상품에 의존
         * 메뉴에 의존
     * 단체 지정 테이블
@@ -182,3 +182,73 @@
     * 단체 지정 테이블
         * 단체 지정 설정 --> 주문 테이블의 단체 지정 값 설정
         * 단체 지정 해제 --> 주문 테이블의 단체 지정 값 초기화
+
+# 4단계 - 멀티 모듈 적용
+
+## 요구사항
+
+* Gradle의 멀티 모듈 개념을 적용해 컨텍스트 간의 독립된 모듈로 만들어본다
+
+초기 계획은 컨텍스트를 생명주기 및 도메인 간 의존성에 따라 총 4가지 모듈로 나누려고 했다
+
+1. 상품 모듈
+    * 상품
+2. 메뉴 모듈
+   * 메뉴
+   * 메뉴 상품
+   * 메뉴 그룹
+3. 주문 모듈
+   * 주문
+   * 주문 메뉴
+   * 주문 상태
+4. 주문 테이블 모듈
+   * 주문 테이블
+   * 단체 지정
+
+
+하지만 구현된 내용으로 의존성 흐름도를 그려본 결과 의존성 cycle이 생기고 있었다
+
+개선 전 의존관계 구성도
+![](img/개선_전_의존관계_구성도.png)
+
+모듈간 의존성 흐름만 간략하게 보면 다음과 같다
+![](img/개선_전_의존관계_흐름도.png)
+
+보이는 것과 같이 `주문` 컨텍스트와 `주문 테이블` 컨텍스트가 서로를 참조하고 있었다
+
+의존성의 가장 상위에 있는 도메인은 `주문` 도메인이다.
+
+따라서 현재 의존성 cycle은 `주문` 에서 `주문 테이블` 로 흘러야 하는 흐름이
+
+`주문 테이블` 에서 `주문` 으로 흐르고 있었기에 발생하였다.
+
+따라서 `주문 테이블` 에서 `주문` 도메인으로 흐르고 있는 의존성이 무엇인지 분석해보니
+총 두 가지의 참조를 발견할 수 있었다
+
+1. 주문 테이블을 빈 테이블로 변경할 경우 `주문의 상태`가 `결제 완료 상태`인지 `체크`한다
+2. 단체 지정을 해제할 때 `주문의 상태`가 `결제 완료 상태`인지 `체크`한다
+
+의존성 cycle의 원인은 `주문의 상태`에 있음을 찾아냈다
+
+따라서 주문 상태에 대한 도메인을 분리하고 
+
+별도의 스키마로 관리하도록 개선을 진행한다
+
+개선 후 의존관계 구성도
+![](img/개선_후_의존관계_구성도.png)
+
+개선 후 의존관계 흐름도
+![](img/개선_후_의존관계_흐름도.png)
+
+개선 이후
+
+`주문 컨텍스트`는 `주문 생성 및 상태 변경` 시 `주문 상태 컨텍스트`에 해당하는 `이벤트를 발행`하고,
+`주문 테이블 컨텍스트`는 `테이블 상태 변경 가능 여부`를 `validation` 하기 위해 `주문 상태 레포지토리를 직접 참조`한다
+
+개선 후 의존관계 흐름도를 보면 
+
+기존에 존재하던 의존성 cycle이 해소된 것을 확인할 수 있다
+
+* 추가 개선사항
+  * 모듈 간 TestFixture 기능 활용
+  * 가독성을 위해 UI 테스트를 RestAssured 로 수행하도록 개선
