+++ kitchenpos-order/src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.order.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
