+++ kitchenpos-order/src/main/java/kitchenpos/order/event/OrderStatusEventPublisher.java
@@ -1,0 +1,19 @@
+package kitchenpos.order.event;
+
+import org.springframework.context.ApplicationEvent;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Component;
+
+@Component
+public class OrderStatusEventPublisher {
+
+    private final ApplicationEventPublisher applicationEventPublisher;
+
+    public OrderStatusEventPublisher(ApplicationEventPublisher applicationEventPublisher) {
+        this.applicationEventPublisher = applicationEventPublisher;
+    }
+
+    public void publish(ApplicationEvent applicationEvent) {
+        applicationEventPublisher.publishEvent(applicationEvent);
+    }
+}
