+++ kitchenpos-order/src/test/java/kitchenpos/OrderApplication.java
@@ -4,8 +4,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-class ApplicationTest {
+public class OrderApplication {
+
     @Test
-    void contextLoads() {
+    public void contextLoads() {
     }
 }
