+++ kitchenpos-orderstatus/src/main/java/kitchenpos/orderstatus/domain/Status.java
@@ -1,0 +1,6 @@
+package kitchenpos.orderstatus.domain;
+
+public enum Status {
+
+    COOKING, MEAL, COMPLETION
+}
