+++ kitchenpos-orderstatus/src/testFixtures/java/kitchenpos/orderstatus/domain/StatusTestFixture.java
@@ -1,0 +1,8 @@
+package kitchenpos.orderstatus.domain;
+
+public class StatusTestFixture {
+
+    public static final Status 조리_상태 = Status.COOKING;
+    public static final Status 식사_상태 = Status.MEAL;
+    public static final Status 계산_완료_상태 = Status.COMPLETION;
+}
