+++ kitchenpos-table/src/main/java/kitchenpos/table/ordertable/domain/OrderTableRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.table.ordertable.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
+    List<OrderTable> findAllByIdIn(List<Long> ids);
+
+    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
+}
